api-server/src/requirements.txt
flask
adafruit-circuitpython-max31855
RPi.GPIO
setuptools
numpy
redis